refactor(iota-traffic-controller): move the traffic simulator into its own module
