+++ crates/iota-e2e-tests/tests/traffic_control_tests.rs
@@ -24,7 +24,7 @@ use iota_sdk_types::{TransactionDigest, UserSignature};
 use iota_swarm_config::network_config_builder::ConfigBuilder;
 use iota_test_transaction_builder::batch_make_transfer_transactions;
 use iota_traffic_controller::{
-    TrafficController, TrafficSim, nodefw_test_server::NodeFwTestServer,
+    TrafficController, nodefw_test_server::NodeFwTestServer, sim::TrafficSim,
 };
 use iota_types::{
     quorum_driver_types::ExecuteTransactionRequestType,
